Allow toggling config persistence
Sometimes, it may not be desirable to always save the configuration
immediately to the startup config. This can be important in the case of
high contention on a switch or many changes happening quickly. It is
faster to apply the changes and not save them in this case. Default to
'true' to keep past behavior.