Trim spaces in ngs_trunk_ports and ngs_physical_networks
This helps configuration readability and avoids surprises.  Currently, any
leading or trailing space in the configuration would be included verbatim
in the name of the port or physical network, which could cause confusion
further down the road.

Now both these forms are accepted and equivalent:

    ngs_trunk_ports = port1,Eth 1/30,Po6

    ngs_trunk_ports = port1, Eth 1/30, Po6

and similarly for ngs_physical_networks

Change-Id: If094fdd475290848fc1e7edf63a04997f7518964