Dell OS9: bypass netmiko save_config (#19)
Depending on OS Version, Dnos9 may not support the command:
`copy running-configuration startup-configuration`, instead
abbreviating it as `copy running-config startup-config`. However,
this asks for a confirmation.

Instead, `write memory` does the same thing.

This PR overrides the superclass so that we send `write memory` instead
of calling out to the netmiko implementation.

This should allow us to revert our netmiko patches, and use upstream
directly.